requirements.txt
PyYAML>=3.12
# brew install libgit2
pygit2>=0.26.3
